LibWeb: Move NavigateParams out of LocalNavigable
The arguments of the navigate algorithm do not depend on which process
hosts the target navigable's active document.
